+++ sdk/core/typespec_client_core/src/http/policies/retry/mod.rs
@@ -109,6 +109,9 @@ pub trait RetryPolicy: std::fmt::Debug + Send + Sync {
     /// If `None` is returned, no headers will be checked.
     fn get_retry_headers(&self) -> Option<&RetryHeaders>;
 
+    /// Get the status codes that should trigger retries.
+    fn get_retry_status_codes(&self) -> &[StatusCode];
+
     /// Determine how long before the next retry should be attempted.
     fn sleep_duration(&self, retry_count: u32) -> Duration;
     /// A Future that will wait until the request can be retried.
@@ -127,10 +130,10 @@ pub trait RetryPolicy: std::fmt::Debug + Send + Sync {
     }
 }
 
-/// The status codes where a retry should be attempted.
+/// Default status codes where a retry should be attempted.
 ///
 /// On all other 4xx and 5xx status codes no retry is attempted.
-const RETRY_STATUSES: &[StatusCode] = &[
+const DEFAULT_RETRY_STATUSES: &[StatusCode] = &[
     StatusCode::RequestTimeout,
     StatusCode::TooManyRequests,
     StatusCode::InternalServerError,
@@ -168,7 +171,8 @@ where
             let (last_result, retry_after) = match result {
                 Ok(response) => {
                     let status = response.status();
-                    if !RETRY_STATUSES.contains(&status) {
+                    let retry_status_codes = self.get_retry_status_codes();
+                    if !retry_status_codes.contains(&status) {
                         if status.is_success() {
                             trace!("server returned success status {}", status,);
                         } else {
@@ -242,7 +246,8 @@ mod test {
     use super::*;
     use crate::http::{
         headers::{Headers, RETRY_AFTER},
-        BufResponse, Context, FixedRetryOptions, Method, Request, RetryOptions, Url,
+        BufResponse, Context, ExponentialRetryOptions, FixedRetryOptions, Method, Request,
+        RetryOptions, Url,
     };
     use ::time::macros::datetime;
     use std::sync::{Arc, Mutex};
@@ -343,7 +348,7 @@ mod test {
         let ctx = Context::new();
         let url = Url::parse("http://localhost").unwrap();
 
-        for &status in RETRY_STATUSES {
+        for &status in DEFAULT_RETRY_STATUSES {
             let mut request = Request::new(url.clone(), Method::Get);
             let count = Arc::new(Mutex::new(0));
             let mock = StatusResponder {
@@ -385,5 +390,116 @@ mod test {
             *count.lock().unwrap(),
             "Policy shouldn't retry after receiving a response whose status isn't in RETRY_STATUSES"
         );
+    }
+
+    #[tokio::test]
+    async fn test_with_retry_statuses() {
+        async fn test_custom_retry_statuses(retry_policy: Arc<dyn Policy>) {
+            let ctx = Context::new();
+            let url = Url::parse("http://localhost").unwrap();
+
+            let mut request = Request::new(url.clone(), Method::Get);
+            let count = Arc::new(Mutex::new(0));
+            let next = vec![Arc::new(StatusResponder {
+                request_count: count.clone(),
+                status: StatusCode::Gone,
+            }) as Arc<dyn Policy>];
+
+            let response = retry_policy
+                .send(&ctx, &mut request, &next)
+                .await
+                .expect("Policy should return a response after exhausting retries");
+
+            assert_eq!(response.status(), StatusCode::Gone);
+            assert_eq!(
+                2,
+                *count.lock().unwrap(),
+                "Policy should retry status in specified list"
+            );
+
+            let mut request = Request::new(url.clone(), Method::Get);
+            let count = Arc::new(Mutex::new(0));
+            let next = vec![Arc::new(StatusResponder {
+                request_count: count.clone(),
+                status: StatusCode::TooManyRequests,
+            }) as Arc<dyn Policy>];
+
+            let response = retry_policy
+                .send(&ctx, &mut request, &next)
+                .await
+                .expect("Policy should return a response without retrying");
+
+            assert_eq!(response.status(), StatusCode::TooManyRequests);
+            assert_eq!(
+                1,
+                *count.lock().unwrap(),
+                "Policy should not retry status not in custom retry list"
+            );
+        }
+
+        let statuses = vec![StatusCode::Gone];
+
+        let retry_policy = RetryOptions::fixed(FixedRetryOptions {
+            delay: Duration::nanoseconds(1),
+            max_retries: 1,
+            ..Default::default()
+        })
+        .with_retry_statuses(statuses.clone())
+        .to_policy(RetryHeaders::default());
+        test_custom_retry_statuses(retry_policy).await;
+
+        let retry_policy = RetryOptions::exponential(ExponentialRetryOptions {
+            initial_delay: Duration::nanoseconds(1),
+            max_retries: 1,
+            ..Default::default()
+        })
+        .with_retry_statuses(statuses)
+        .to_policy(RetryHeaders::default());
+        test_custom_retry_statuses(retry_policy).await;
+    }
+
+    #[tokio::test]
+    async fn test_empty_retry_statuses_disables_status_retries() {
+        async fn test_no_retries_for_default_statuses(retry_policy: Arc<dyn Policy>) {
+            let ctx = Context::new();
+            let url = Url::parse("http://localhost").unwrap();
+            for &status in DEFAULT_RETRY_STATUSES {
+                let mut request = Request::new(url.clone(), Method::Get);
+                let count = Arc::new(Mutex::new(0));
+                let mock = StatusResponder {
+                    request_count: count.clone(),
+                    status,
+                };
+                let next = vec![Arc::new(mock) as Arc<dyn Policy>];
+                let result = retry_policy
+                    .send(&ctx, &mut request, &next)
+                    .await
+                    .expect("Policy should return without retrying");
+                assert_eq!(result.status(), status);
+                assert_eq!(
+                    1,
+                    *count.lock().unwrap(),
+                    "Policy should not retry {status} when given an empty list of retry statuses"
+                );
+            }
+        }
+
+        let retry_policy = RetryOptions::fixed(FixedRetryOptions {
+            delay: Duration::nanoseconds(1),
+            max_retries: 1,
+            ..Default::default()
+        })
+        .with_retry_statuses(vec![])
+        .to_policy(RetryHeaders::default());
+        test_no_retries_for_default_statuses(retry_policy).await;
+
+        let retry_policy = RetryOptions::exponential(ExponentialRetryOptions {
+            initial_delay: Duration::nanoseconds(1),
+            max_retries: 1,
+            ..Default::default()
+        })
+        .with_retry_statuses(vec![])
+        .to_policy(RetryHeaders::default());
+        test_no_retries_for_default_statuses(retry_policy).await;
     }
 }